requirements.txt
ipython
ipdb
paho-mqtt
pytest
